Added function to sort coffee shop list by distance and refactoring
- Renamed DistanceCalculator to DistanceUtility
- Added another test and function to DistanceUtility that sorts coffee shops from a list according to their distance to a specified location
